+++ ruma-client-api/CHANGELOG.md
@@ -65,6 +65,8 @@ Breaking changes:
 * Move `user: UserInfo` in `r0::session::login::Request` to `identifier: UserIdentifier` in
   `r0::session::login::LoginInfo::Password`
   * `r0::session::login::Request::new` takes only `login_info: LoginInfo` as a param
+* Change `ruma_events::AnyEvent` to `ruma_events::AnySyncRoomEvent` in
+  `push::get_notifications::Notification`
 
 Improvements: